+++ avs-api/pom.xml
@@ -36,14 +36,14 @@
         </dependency>
 
         <dependency>
-            <groupId>com.google.dagger</groupId>
-            <artifactId>dagger</artifactId>
-            <version>${dagger.version}</version>
-        </dependency>
-        <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
             <version>${jackson.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.dagger</groupId>
+            <artifactId>dagger</artifactId>
+            <version>${dagger.version}</version>
         </dependency>
         <dependency>
             <groupId>io.undertow</groupId>
@@ -65,13 +65,19 @@
         </dependency>
         <dependency>
             <groupId>org.example.age</groupId>
-            <artifactId>testing-server</artifactId>
+            <artifactId>test-server</artifactId>
             <version>${revision}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.example.age</groupId>
-            <artifactId>testing-service</artifactId>
+            <artifactId>test-service</artifactId>
+            <version>${revision}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.example.age</groupId>
+            <artifactId>testing-server</artifactId>
             <version>${revision}</version>
             <scope>test</scope>
         </dependency>
