+++ avs-api/src/test/java/org/example/age/avs/api/AvsApiTest.java
@@ -16,15 +16,15 @@ import java.util.Map;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import okhttp3.Response;
-import org.example.age.avs.api.testing.StubAvsService;
+import org.example.age.avs.api.test.StubAvsService;
 import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
 import org.example.age.data.DataMapper;
 import org.example.age.data.SecureId;
 import org.example.age.data.certificate.VerificationSession;
+import org.example.age.test.server.undertow.TestUndertowModule;
+import org.example.age.test.service.data.TestAccountIdExtractorModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestUndertowServer;
-import org.example.age.testing.server.undertow.TestUndertowModule;
-import org.example.age.testing.service.data.TestAccountIdExtractorModule;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
