+++ common/src/test/java/org/example/age/common/avs/api/AvsApiHttpHandlerTest.java
@@ -21,11 +21,11 @@ import org.example.age.common.avs.store.InMemoryRegisteredSiteConfigStoreModule;
 import org.example.age.common.avs.store.InMemoryVerifiedUserStoreModule;
 import org.example.age.common.base.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
+import org.example.age.test.server.undertow.TestUndertowModule;
+import org.example.age.test.service.data.TestAccountIdExtractorModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.crypto.TestKeys;
 import org.example.age.testing.server.TestUndertowServer;
-import org.example.age.testing.server.undertow.TestUndertowModule;
-import org.example.age.testing.service.data.TestAccountIdExtractorModule;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
