+++ common/src/test/java/org/example/age/common/site/api/SiteApiHttpHandlerTest.java
@@ -28,11 +28,11 @@ import org.example.age.common.site.config.SiteConfig;
 import org.example.age.common.site.store.InMemoryVerificationStoreModule;
 import org.example.age.data.SecureId;
 import org.example.age.data.certificate.VerificationSession;
+import org.example.age.test.server.undertow.TestUndertowModule;
+import org.example.age.test.service.data.TestAccountIdExtractorModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.crypto.TestKeys;
 import org.example.age.testing.server.TestUndertowServer;
-import org.example.age.testing.server.undertow.TestUndertowModule;
-import org.example.age.testing.service.data.TestAccountIdExtractorModule;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
