+++ demo/pom.xml
@@ -39,7 +39,7 @@
         <!-- test -->
         <dependency>
             <groupId>org.example.age</groupId>
-            <artifactId>testing</artifactId>
+            <artifactId>testing-data</artifactId>
             <version>${revision}</version>
             <scope>test</scope>
         </dependency>
