+++ pom.xml
@@ -12,7 +12,7 @@
     <modules>
         <!-- data -->
         <module>data</module>
-        <module>testing</module>
+        <module>testing-data</module>
 
         <!-- api types -->
         <module>api</module>
@@ -29,12 +29,12 @@
         <module>common-service</module>
         <module>common</module>
         <module>verification-poc</module>
-        <module>testing-service</module>
+        <module>test-service</module>
 
         <!-- server -->
         <module>common-server</module>
-        <module>common-server-test</module> <!-- Break a circular dependency in the test scope. -->
         <module>adult</module>
+        <module>test-server</module>
         <module>testing-server</module>
 
         <!-- demo -->
