+++ test-service/pom.xml
@@ -9,44 +9,43 @@
         <artifactId>age-verification</artifactId>
         <version>${revision}</version>
     </parent>
-    <artifactId>data</artifactId>
+    <artifactId>test-service</artifactId>
     <packaging>jar</packaging>
 
     <dependencies>
         <!-- main -->
         <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>${jackson.version}</version>
+            <groupId>org.example.age</groupId>
+            <artifactId>api</artifactId>
+            <version>${revision}</version>
         </dependency>
         <dependency>
-            <groupId>com.fasterxml.jackson.datatype</groupId>
-            <artifactId>jackson-datatype-guava</artifactId>
-            <version>${jackson.version}</version>
+            <groupId>org.example.age</groupId>
+            <artifactId>common-api</artifactId>
+            <version>${revision}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.google.dagger</groupId>
+            <artifactId>dagger</artifactId>
+            <version>${dagger.version}</version>
         </dependency>
         <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>${guava.version}</version>
+            <groupId>io.undertow</groupId>
+            <artifactId>undertow-core</artifactId>
+            <version>${undertow.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.immutables</groupId>
-            <artifactId>value-annotations</artifactId>
-            <version>${immutables.version}</version>
+            <groupId>javax.inject</groupId>
+            <artifactId>javax.inject</artifactId>
+            <version>1</version>
         </dependency>
 
         <!-- test -->
         <dependency>
             <groupId>org.example.age</groupId>
-            <artifactId>testing-data</artifactId>
+            <artifactId>testing-api</artifactId>
             <version>${revision}</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava-testlib</artifactId>
-            <version>${guava.version}</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -60,9 +59,9 @@
                 <configuration>
                     <annotationProcessorPaths>
                         <dependency>
-                            <groupId>org.immutables</groupId>
-                            <artifactId>value</artifactId>
-                            <version>${immutables.version}</version>
+                            <groupId>com.google.dagger</groupId>
+                            <artifactId>dagger-compiler</artifactId>
+                            <version>${dagger.version}</version>
                         </dependency>
                     </annotationProcessorPaths>
                 </configuration>
