+++ testing-data/pom.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.example.age</groupId>
+        <artifactId>age-verification</artifactId>
+        <version>${revision}</version>
+    </parent>
+    <artifactId>testing-data</artifactId>
+    <packaging>jar</packaging>
+</project>
