--- testing-service/src/main/java/org/example/age/testing/service/data/TestAccountIdExtractorModule.java
@@ -1,16 +1,0 @@
-package org.example.age.testing.service.data;
-
-import dagger.Binds;
-import dagger.Module;
-import org.example.age.common.api.data.AccountIdExtractor;
-
-/**
- * Dagger module that publishes a binding for {@link AccountIdExtractor},
- * which reads the custom {@code Account-Id} header, or sends a 401 error.
- */
-@Module
-public interface TestAccountIdExtractorModule {
-
-    @Binds
-    AccountIdExtractor bindAccountIdExtractor(TestAccountIdExtractor impl);
-}
